// src/models/MimeType.ts
export enum MimeType {
    MP3 = "audio/mpeg",
    OGG = "audio/ogg",
    MP4 = "video/mp4",
    AVI = "video/avi",
    MOV = "video/quicktime",
    WEBM = "video/webm",
    MPEG = "video/mpeg",
    PNG = "image/png",
    JPEG = "image/jpeg",
    JPG = "image/jpg",
    GIF = "image/gif",
    BMP = "image/bmp",
    SVG = "image/svg+xml",
    WEBP = "image/webp",
    SWF = "application/x-shockwave-flash",
    UNKNOWN = "unknown",
}